Fix traymenu update issue
