+++ Sources/AppBundle/TrayMenuModel.swift
@@ -1,5 +1,11 @@
 import AppKit
 import Common
+import SwiftUI
+
+struct WorkspaceButton: Identifiable {
+    let id = UUID()
+    let button: Button<Toggle<Text>>
+}
 
 public class TrayMenuModel: ObservableObject {
     public static let shared = TrayMenuModel()
@@ -9,6 +15,7 @@ public class TrayMenuModel: ObservableObject {
     @Published var trayText: String = ""
     /// Is "layouting" enabled
     @Published var isEnabled: Bool = true
+    @Published var workspaceStatus: [WorkspaceButton] = []
 }
 
 func updateTrayText() {
@@ -19,4 +26,17 @@ func updateTrayText() {
             ($0.activeWorkspace == focus.workspace && sortedMonitors.count > 1 ? "*" : "") + $0.activeWorkspace.name
         }
         .joined(separator: " │ ")
+    TrayMenuModel.shared.workspaceStatus = Workspace.all.map { (workspace: Workspace) in
+        WorkspaceButton(button: Button {
+            refreshSession { _ = workspace.focusWorkspace() }
+        } label: {
+            Toggle(isOn: workspace == focus.workspace
+                ? Binding(get: { true }, set: { _, _ in })
+                : Binding(get: { false }, set: { _, _ in }))
+            {
+                let monitor = workspace.isVisible || !workspace.isEffectivelyEmpty ? " - \(workspace.workspaceMonitor.name)" : ""
+                Text(workspace.name + monitor).font(.system(.body, design: .monospaced))
+            }
+        })
+    }
 }
